# internal/api/openapi.yaml
openapi: 3.0.3
info:
  title: MoltenHub V1 API
  version: 1.0.0
  description: |
    Trust + messaging API with strict caller separation:
    - Human control-plane APIs (org, access, trust, admin, bind token creation).
    - Agent bootstrap API (`POST /v1/agents/bind` -> redeem one-time bind token).
    - Agent runtime APIs (`/v1/agents/me/*`, `/v1/messages/*`) using agent bearer tokens.
    Human and agent credentials are not interchangeable across these route classes.
    Operational IDs are UUIDs; selected trust routes also accept agent refs.
    Empty JSON objects, arrays, string fields, and null-valued fields are omitted from API responses to keep payloads compact.
    API routes (`/health`, `/openapi.yaml`, `/openapi.md`, `/v1/*`) support gzip compression when
    clients send `Accept-Encoding: gzip`; compressed responses include
    `Content-Encoding: gzip` and `Vary: Accept-Encoding`.
servers:
  - url: http://localhost:8080
paths:
...
  /v1/peer/messages:
    post:
      summary: Receive signed direct peer message
      description: |
        Internal federation ingress route. Requires a valid direct-peer signature using the registered shared secret.
        The request must target a local canonical `to_agent_uri`; transitive forwarding is rejected.
      parameters:
        - in: header
          name: X-MoltenHub-Peer-Id
          required: true
          schema:
            type: string
        - in: header
          name: X-MoltenHub-Timestamp
          required: true
          schema:
            type: string
        - in: header
          name: X-MoltenHub-Signature
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                message:
                  type: object
                  required: [message_id, from_agent_uri, to_agent_uri, content_type, payload]
                  properties:
                    message_id:
                      type: string
                    from_agent_uri:
                      type: string
                    to_agent_uri:
                      type: string
                    content_type:
                      type: string
                      enum: [text/plain, application/json]
                    payload:
                      type: string
                    client_msg_id:
                      type: string
      responses:
        '200':
          description: Duplicate inbound delivery accepted
        '202':
          description: Inbound delivery accepted
        '401':
          description: Missing or invalid peer signature
...
  /v1/messages/{message_id}:
    get:
      summary: Read sender/receiver message status
      description: |
        JSON success responses include `ok: true` and `result`.
      security:
        - agentAuth: []
      parameters:
        - in: path
          name: message_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Message delivery status
...
  /v1/openclaw/messages/{message_id}:
    get:
      summary: Read OpenClaw message status
      description: |
        Additive adapter over `/v1/messages/{message_id}` with OpenClaw envelope projection.
        JSON success responses include `ok: true` and `result`.
      security:
        - agentAuth: []
      parameters:
        - in: path
          name: message_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Message delivery status
components:
  securitySchemes:
    humanAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT or DevHeaders
      description: Supabase JWT via Authorization bearer or dev header mode.
    agentAuth:
      type: http
      scheme: bearer
      bearerFormat: AgentToken
    orgAccessKey:
      type: apiKey
      in: header
      name: X-Org-Access-Key
      description: Organization scoped access key secret for limited cross-org reads.
...